Initial commit.
This commit contains essentially essentially verbatim the contents of the
"X-Ares ∂ ƒ" folder as released in "Ares GNU GPL Source".  The "Project
Libraries" folder has been excluded on the assumption that the files therein
mostly pertain to outdated Classic system calls and they will need to be
replaced wholesale.

The COPYING and README files have been added; the original source release
contained copies of these in RTF format but plain-text is more common in
open-source projects.