Add StringNumerics.{hpp,cpp}.
These files were part of the "Project Libraries" folder of the original
distribution; no license was given, so the GPL header has been added to both.
There was a C++ source file, but I have opted to use the C source file and
rename it, on the assumption that if there are any differences, then it would
be better to use the one Ares used as a starting point.