Decouple clock from real clock.
If the game has not yet started, TickTime() is incremented by one each time it
is called; following the start of the game, it is incremented by one for each
game tick.

This is necessary because the sound code makes decisions about persistence
based on TickTime(), and so using the real clock caused too many sounds to be
played at once.