Fix warnings exposed by new compiler settings.
This change also adds two new enums.  The hope was that gcc would be able to
use these definitions to determine that a switch's list of cases was
exhaustive, and that this would eliminate compiler warnings.

Unfortunately, either gcc is not able to use this information for control flow,
or it's being defensive about integers that are cast to invalid enum values
(which we do in some places).  However, even if it doesn't solve the problem,
it was something that ought to be done eventually anyway, so it's being
committed as is.