Fix bug in PixelInSprite_IsOutside().
The original implementation of the function considered a pixel to be inside if
it was in the first or last row or column, or if any of the physical pixels
(i.e. offset according to hmap and vmap) at (x +/-, y +/- 1) were empty.

The problem with this algorithm is that it was checking if the pixel was in the
last row or column according to its logical (unscaled) location, rather than
its physical location.  That meant that pixels in the last column were being
compared against pixels early in the next row, and pixels in the last row were
being compared against unallocated space.