Fix bug with visible starmap edges during Mission Analysis.
The starmap image is 3 pixels taller than the area in which the tactical map is
shown.  The starmap image would be copied to the real GWorld; later screens
would only copy the portion of the off GWorld in which the tactical map was
displayed, leaving the remaining three rows of pixels from the starmap visible.