Replace DoMainScreenInterface with an EventListener.
There are two problems unresolved by this commit: first, when a button is
clicked, the method that is called runs its own event loop, so the main screen
is not entirely handled in terms of the EventListener interface.  Second, the
'About Ares' scrolling text does not display correctly due to its use of some
more advanced features.