Add libsfz and libpng submodules; use gyp to build.
libsfz is as of yet unused; the Antares versions of these classes will be
retired in a follow-up commit.  libpng has been updated to v1.4.0, the most
recent stable release, with a few small associated changes.

Since the new build system is slightly stricter with its warnings, changes have
been made to a few classes in order to silence them, mostly to add virtual
destructors.  Also, scripts/test has been updated to use the new build system
when updating, and to run the binaries from their new location in build/.