Always dequeue events before current time.
At times when the interval between timers was small (read: during GamePlay), we
previously had a heuristic to always dequeue an event that had been seen during
the previous call to antares_event_translator_enqueue().  This prevented event
starvation, because we were now always guaranteed to make progress on the event
queue.

However, it wasn't enough.  Only dequeueing a single event meant that, during
times when the interval between events was small (read: moving the mouse or
mashing on the keyboard) the queue would grow faster than it was drained,
resulting in noticeable input lag.

Now, we  always drain the event queue of events that occured before the current
time.  This ensures that the queue does not grow when there are lots of events.