Split libAntares into per-dir libs.
When `ar` adds a file to an archive, it ignores all but the final path
component of the added file.  As a result, there were some collisions (e.g.
sound/driver.hpp and video/driver.hpp), resulting in problems when attmepting
to use `gdb` on the resulting binary.  Having one stlib per directory prevents
this from happening.