Prevent crash when losing level.
Previously, losing the level resulted in _typed_chars and _state being
initialized to 0 and DONE, and the latter meant _typed_chars would never
be incremented.  Then, because _typed_chars was 0, there were never any
draw_char() calls.  With draw_range(), we need to prevent it from being
called when _score is null.