Exclude sprites based on right edge of viewport
The exclusion code from before seems wrong. The idea behind
`viewport.left - kSpriteMaxSize` as the left bound is that any sprite
left of that cannot extend into the viewport. But for the right side,
you’d expect `viewport.right + kSpriteMaxSize`; i.e., that a sprite to
the right of that cannot extend into the right side of the viewport.